(PC-9515) rename useHomeAlgoliaModules => useHomeModules
